Add files via upload
This is one of my projects for the autonomous embedded system in a vehicle, and this project only uses the servo motor, distance sensor, and all the I/O sensors available in the Grove Beginner kit for Arduino.  This is just a prototype for smaller tasks in an automobile that can be performed using smaller sensors.